Linked-list implementation now not considered public for Subscription
Adjusts the public API of the Subscription struct to work even
in case we switch the implementation away from a linked list.
Mainly, the index is no longer a public property since it might
not be updatable in the structure right away.  It might require
look-up, e.g. if subscription data inside Promise were stored
in a B-Tree, for example.