No longer have ValueTask<PromiseResult>.GetResult throw on cancellation
Throwing exceptions makes it difficult to dispose the subscription,
since the structure that is the target of the "using" statement
has to be returned (normally) for that to happen.

Instead, signal cancellation with a status flag, PromiseStatus,
and then PromiseResult can consult that flag, and throw exceptions
when unavailable properties get accessed.