Rename TakeJob to TryTakeItem and change to a more conventional signature
The "conventional" signature has a boolean result value and
the item as an "out" parameter.  This reduces friction when
adapting to the ChannelReader<T>.TryRead API.

The previous signature we had worked well when the item type
is not generic.  Now that it is, we should not disallow
putting null as an item even though in most applications that
would not be useful.