Use classes instead of delegates to represent RPC callback internally
Classes allow any number of (internal) parameters to be
captured efficiently.  With delegates we have to
store and pass a type-erased "object" parameter,
and to extend to a new parameter we would have to
wrap the parameters in a (tuple) object.

Also define a delegate for user-defined RPC functions,
so extra arguments can be added easily.  For now we
add CancellationToken although cancellation is not
implemented yet.