No longer use delegates to represent "promise jobs"
Delegates (functions) are obviously not possible to send to
remote hosts for distributed work.

Instead, describe job inputs purely in terms of serializable
data, and optional command strings and hint objects.

Thus "PromiseJobFunction" becomes simply "PromiseJob",
in this revision.

It had been envisioned that the delegate, when run locally,
would in turn invoke other APIs to send the job to a
remote host.  However, this additional layer of indirection
makes the API complex.  Instead, now we will just require
the programmatic user of our system implement
IJobWorker<PromiseJob, PromiseData> to run the job.