Have WorkerDistribution be instantiated outside of JobSchedulingSystem
It should be a separate dependency-injected component,
as there will be options that the user will want to control.

The mock DummyWorker class that was private to JobSchedulingSystem
is now moved into Server.Program.  It should not belong in
JobSchedulingSystem since that exists in the core library which
should be free of demo code and mocks.