Remove JobsServerConfiguration class
This class grouped the services to configure ASP.NET Core endpoints
for the promises system, but those configuration methods can retrieve
services from IEndpointRouteBuilder by themselves, so the caller
does not need to pass them.

We do not expect the same server to be running multiple independent
instances of PromiseStorage and PathsDirectory, but if they do,
they will need to scope those services in ASP.NET Core's dependency
injection framework.