Return boolean from IncrementalAsyncList<T>.TryGetMember to mark end of list
Before the method returned ValueTask<T?>.

Despite T having a constraint of "notnull", the C# compiler seems
to refuse to turn it into Nullable<T> when T is a value type,
even though the documentation seems to say that is what should
be done.  To work around this difficulty, we re-design the
method to return ValueTask<(T, bool)> instead.

At the same time, the exception stored with completion will
not be thrown out of this method any longer, which also
makes it more efficient.  The client must query the new
"Exception" property explicitly.