Mark PromiseList result as transient if any element is transient
This would matter in the following situation:

A job whose promise was put into a PromiseList and gets
administratively cancelled would report OperationCanceledException,
but the macro job would not get that OperationCanceledException
and would complete the PromiseList without any exception.

But still, if the macro job is requested again, it should
restart itself, just as for a micro job that was administratively
cancelled.