Remove ScheduledJob structure
This structure had been first designed to wrap SharedFuture<TInput, TOutput>.

Since then the latter has been abstracted to an interface
ILaunchableJob<TInput, TOutput>.  Furthermore, the only other
member of ScheduledJob, of type ISchedulingAccount, has never
been used.  So we can remove the complexity of this wrapper
structure.

The reference to ISchedulingAccount had been intended to
inform the job de-queuing algorithm of which (client) queue a
job comes from.  However, ISchedulingAccount would not be
an adequate representation of the queue anyway.  If we
need this information, it is better to have the queue,
that is the implementation of SchedulingFlow<T>, stamp
its outputs on demand.