Re-design Web API endpoint for cancelling jobs
There should be a dedicated endpoint accept a HTTP DELETE request
so that authorization policies can be easily applied.

Before, the endpoint was multiplexed between several possible
actions with HTTP POST, with a view of being able to use it
from an HTML form (which only supports GET and POST).
Now we preclude that possibility which had always been
theoretical: a Web browser client must use JavaScript
if it wants to call this ReST API.