Add "mean wait time" to MockPricingInput and randomize waiting time
Before, the test wait time was recorded in PromisedWork.InitialWait.
Now, there is a separate member in MockPricingInput so that
we can test the adjustment of queue charge time when the two
members differ.

With this change, the unit tests now have the opportunity to
specify the waiting time to be minimal, to make them run faster.

Furthermore, randomize the waiting time within MockPricingWorker.