Remove CopyLocalLockFilesAssemblies = false from MSBuild properties
This option has stopped working after upgrading to the latest
.NET 6 packages.  Formerly, NuGet dependencies were consumable
directly from the global cache of NuGet-downloaded files.

The symptom is "FileNotFound" exceptions on assemblies like
Microsoft.TestPlatform.Abstractions, when trying to run
the tests.

The error messages seem to indicate mismatched versions of
certain assemblies included from FrameworkReference.  This
bug in the .NET SDK may be relevant:

https://github.com/dotnet/sdk/issues/22986