Specify "ui" sub-path explicitly in Blazor server pages
Using <base href="...">, described as option 1 in

  https://docs.microsoft.com/en-us/aspnet/core/blazor/host-and-deploy/

fails when a Razor Components Library (RCL) needs to deploy
JavaScript initializers:

  https://docs.microsoft.com/en-us/aspnet/core/blazor/javascript-interoperability/

The Web browser ends up requesting the JavaScript initializer
script at the URL:

  https://‹host›/server/ui/_content/Hearty.Server.WebUi/…

when the server actually installs the script at the URL:

  https://‹host›/server/_content/Hearty.Server.WebUi/…

These paths seem to be hard-coded to the ASP.NET Core
framework, or at least are very hard to change.  So now
we set the pages up in the manual but foolproof way:
each Blazor page has the "/ui/" prefix specified for
@page, and all links written in Razor and Blazor pages
all have the "ui/" prefix.

Fortunately, the path prefix for the whole job server,
can still be set by "UseStrictPathBase" and gets
propagated to the Web browser through:

  <base href="~/" />

which is substituted by Razor before it is sent
to the client.