Translate termination exception in PromiseList to PromiseExceptionalData
Translate the Exception object on receiving it and store it as
an instance member, so that it can be readily serialized.

Before, PromiseExceptionalData was created each time the client
requests the payload of PromiseList.

IncrementalAsyncList is changed to accept arbitrary objects
as the termination status, and no longer treats Exceptions
specially, which simplifies the code!