Support optional "prologue row" in Blazor Grid control
The "prologue row" is rendered below the header row, for now.

It is intended to be rendered above, but that screws up column
resizing because the table layout for the grid is "fixed",
and that means the first row of the table determines
the width of the columns.

Moving the prologue row above the header row needs a more
involved change: the width should be controlled by a HTML
col element instead of directly on the th cell.