Support LINQ expressions in PropertyColumn for Blazor grid control
The original code (which had been copied from a Microsoft demo)
supported LINQ expressions already.  For efficiency and simplicity
we had removed that support.

Now we add it back, although the feature remains optional where
it was mandatory in the original code.  Items sources for the
grid control which do not come from a database, i.e. does
not implement IQueryable<TGridItem>, can continue to supply plain
functions for the property projection.