Upgrade to FasterKV version 2.0.11
v2.x is a new major version with API chnages.

This revision minimally adapts the existing code to compile
and run, but does not attempt to use the new API fully.

The author is hoping this upgrade will fix some reliability
bugs --- though there is also a risk of regressing as well
from all of Faster KV's internal architectural changes.