Use upsert instead of RMW to implement FasterKV-based promise storage
RMW had been used to work around a limitation of Faster KV in its
earlier major version.  It seems there might be a reliability problem
with RMW.  We switch to upsert and see if that helps.

This change is experimental, and if it works, the code will need
some cleaning up.