Re-design object cache in FasterDbPromiseStorage to avoid race conditions
It turns out that storing GCHandle directly inside ConcurrentDictionary
does not work.  ConcurrentDictionary guarantees there are no torn reads
when there are concurrent writes but that is about the only guarantee.
It is possible, for instance, that even when an entry is removed,
another thread can be concurrently using the same entry that has just
been read before.  So operations like GCHandle.Free become dangerous,
particularly as the freed handles can be re-used quickly by the .NET
run-time.

The problems showed up in stress tests of an application by the author.

To fix the race conditions, the cache of live objects must be re-designed
to store references to objects, i.e. we use the reference type
WeakReference<Promise> instead of GCHandle.  The memory of the
referenced object can then be used (under the hood by the .NET run-time)
for any required inter-thread synchronization.