Fix a rare ABA race condition in weak-reference object pooling
In summary, when there is pooling of the weak reference objects,
it is not allowed for multiple threads to try to remove entries
from the cache of live objects.  When attempting to remove an
existing entry, it is always verified for the desired value to
remove, but that only works through comparison of references.
If the weak reference object is re-used from the pool,
that verification could potentially result in a false positive.

In detail, the following erroneous sequence of operations
could happen, when attempting to concurrently remove cache entries,
where "obj" is seen by Thread 1 and 2 as expired:

Thread 1         Thread 2             Thread 3
                   1. Read obj
  2. Remove obj
  3. Pool obj
                                        4. Take obj from pool
                                        5. Store obj into entry
                   6. Remove obj (BAD)

The ABA race condition could be addressed by incrementing a
counter every time a weak reference object gets re-used, and
then storing a snapshot of the counter inside the cache entry.

Unfortunately, such ABA tagging makes the code more complicated,
and also slower for the common case.  So this revision
changes the code to give up on trying to aggressively recycle
the weak reference objects.

This problem was identified by the author's purely
logical reasoning; it would be very difficult to trigger
in practice.