Need to re-create the HttpRequestMessage when posting jobs on retry
The preceding revisions to make HeartyHttpClient allow retries
do not work without this change.

It is a "gotcha" with HttpRequestMessage:

  https://github.com/dotnet/runtime/issues/24099

The request message can be mutated by the chain of handlers
within HttpClient, so it is not allowed (by Microsoft) to
re-use the same instance to make another HTTP request.