Remove PathsDirectory and associated infrastructure
The whole idea of retrieving promises by "paths" turned out
to be functionality that does not belong in the core,
and can be implemented in better ways.

For instance, the segments in the path could be stored as
separate database columns (e.g. in SQLite).  The Hearty
framework should not impose a specific representation of names
or paths of promises, but leave those details to the
individual applications.